+++ api/src/main/java/org/allaymc/api/item/ItemStack.java
@@ -1,12 +1,9 @@
 package org.allaymc.api.item;
 
 import org.allaymc.api.item.component.ItemBaseComponent;
-import org.allaymc.api.item.component.data.ItemDataComponent;
 
 /**
  * @author daoge_cmd
  */
-public interface ItemStack extends
-        ItemBaseComponent,
-        ItemDataComponent {
+public interface ItemStack extends ItemBaseComponent {
 }