+++ server/src/main/java/org/allaymc/server/item/impl/ItemStackImpl.java
@@ -4,7 +4,6 @@ import lombok.experimental.Delegate;
 import org.allaymc.api.component.interfaces.Component;
 import org.allaymc.api.item.ItemStack;
 import org.allaymc.api.item.component.ItemBaseComponent;
-import org.allaymc.api.item.component.data.ItemDataComponent;
 import org.allaymc.api.item.initinfo.ItemStackInitInfo;
 import org.allaymc.server.component.ComponentClass;
 import org.allaymc.server.component.interfaces.ComponentProvider;
@@ -16,7 +15,6 @@ import java.util.List;
  */
 public abstract class ItemStackImpl extends ComponentClass implements ItemStack {
     protected ItemBaseComponent baseComponent;
-    protected ItemDataComponent dataComponent;
 
     public ItemStackImpl(ItemStackInitInfo initInfo, List<ComponentProvider<? extends Component>> componentProviders) {
         super(initInfo, componentProviders);
@@ -25,10 +23,5 @@ public abstract class ItemStackImpl extends ComponentClass implements ItemStack 
     @Delegate
     protected ItemBaseComponent getBaseComponent() {
         return baseComponent;
-    }
-
-    @Delegate
-    protected ItemDataComponent getDataComponent() {
-        return dataComponent;
     }
 }